Compat: Windows

+++ src/corn/media/interface_impl.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <bgfx/bgfx.h>
-#include <GLFW/glfw3.h>
 #if BX_PLATFORM_LINUX
 #define GLFW_EXPOSE_NATIVE_X11
 #elif BX_PLATFORM_WINDOWS
@@ -9,6 +8,7 @@
 #elif BX_PLATFORM_OSX
 #define GLFW_EXPOSE_NATIVE_COCOA
 #endif
+#include <GLFW/glfw3.h>
 #include <GLFW/glfw3native.h>
 #include "../render/shader.h"
 
@@ -17,7 +17,7 @@ namespace corn {
     public:
         GLFWwindow* window;
         bgfx::ViewId viewID;
-        int width, height, fwidth, fheight;
+        uint16_t width, height, fwidth, fheight;
 
         Shader polygonShader;
